+++ mayan/apps/sources/literals.py
@@ -16,3 +16,5 @@ DEFAULT_SOURCES_CACHE_STORAGE_BACKEND_ARGUMENTS = {
 DEFAULT_SOURCES_LOCK_EXPIRE = 600
 
 STORAGE_NAME_SOURCE_CACHE_FOLDER = 'sources__source_cache'
+
+SOURCE_ACTION_EXECUTE_TASK_PATH = 'mayan.apps.sources.tasks.task_source_process_document'