+++ mayan/apps/sources/source_backends/mixins.py
@@ -11,6 +11,7 @@ from mayan.apps.documents.models.document_type_models import DocumentType
 from mayan.apps.documents.tasks import task_document_file_upload
 
 from ..classes import DocumentCreateWizardStep
+from ..literals import SOURCE_ACTION_EXECUTE_TASK_PATH
 from ..tasks import task_process_document_upload
 
 from .literals import (
@@ -317,7 +318,7 @@ class SourceBackendPeriodicMixin:
                 obj={'source_id': self.model_instance_id}
             ),
             name=self.get_periodic_task_name(),
-            task='mayan.apps.sources.tasks.task_source_process_document'
+            task=SOURCE_ACTION_EXECUTE_TASK_PATH
         )
 
     def delete(self):