+++ HISTORY.rst
@@ -9,6 +9,7 @@
   of files if ``MAYAN_COMMON_DISABLE_LOCAL_STORAGE`` is set to any truthy
   value (True, true, T, t, Yes, yes, Y, y, 1).
 - Backport periodic task import checking.
+- Backport source periodic task changes.
 
 4.4.8 (2023-07-15)
 ==================